Fix Windows idempotency issues & add proxy support
The DisplayName value was incorrect in the window_package call.
This has been corrected, however in doing so we have identified
a bug in the installer which may force us to have to check for
two different DisplayName values in the future.

Added support to the varfile for proxies as this is a commonly
used feature.