Make linux correctly idempotent.
The value for ['rapid7']['product'] correctly has a trailing
space at the end to facilitate idempotency on Windows. However,
this change inadvertently broke idempotency on Linux as the
not_if gaurd was looking for a path containing said trailing
space. The included chefspec test for this issue helped identify
this issue, however broke after fixing the issue.