Fix travis builds.
This switches the travis builds to use the EC2 driver instead of the
docker driver. This was done primarily because the current state of the
cookbook depends on the service resource which technically does not
function with most docker containers.

Additionally, I have added support for testing Windows systems in EC2
with this change to help validate pull requests against Windows.