Add missing values to .kitchen.yml
I had some values which were set in my local .kitchen.yml file and were
being merged with the .kitchen.local.yml file which caused it all to
work correctly for me. As result the behavior was correctly broken in
travis as it had missing parameters such as the driver name and other
important values. This should fix it all for sure now.