Switch to kitchen-sync transport and pin inspec gem.
Noah Kantrowitz (coderanger) was kind enough to fix the bug in the
inspec verifier today, and it was quickly merged.

https://github.com/chef/kitchen-inspec/issues/31

The fix has not been officially released, and as such we need to pin
the github ref until it is released. Moreover, this switches the
transport to kitchen-sync as I know it works propely with that gem.