Fix init script to prevent Chef restarting Nexpose
The init scripts did not support the `status` command which made it
impossible for Chef to determine if it should start the service or not.
This fix has been deployed into the nexpose code base, but has not yet
been released. As such this patch will fix the issue for customers
currently experiencing the issue without having to update or restart.