Fetch nexpose installer artifact over HTTPS instead of HTTP
After fetching this URI the recipe runs it as root.  Short of
verifying the integrity of the installer in some way, the installer
should be downloaded over HTTPS at the least.

Note that there's a `checksum` directive in the `remote_file`, but
that doesn't do what you might think it does: instead of verifying
that the downloaded artifact has the expected checksum, it just uses
the checksum as a hint to tell whether to re-download the file.